pyproject.toml
[tool.poetry]
name = "ukrfans-discord"
version = "0.0.1"
description = "Discord bot for Ukrfans Discord server."
authors = ["V1def"]
license = "AGPL"

[tool.poetry.dependencies]
python = "^3.8"
disnake = "^2.1.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
